Implement Singleton pattern - to only load models on startup
