+++ backend/pipelines/steps/embed_clusters.py
@@ -1,20 +1,7 @@
 import pickle
-from backend.app import database
-from backend.app.services.tf_idf_embedder import TFIDFEmbeddingService
-from backend.app.services.sbert_embedder import SBERTEmbeddingService
+from backend.app.services.tf_idf_embedder import load_vectorizer
+from backend.app.services.sbert_embedder import get_sbert_service
 from backend.app import models
-
-def load_vectorizer(path="tfidf_vectorizer.pkl"):
-    """
-    Load the previously fitted TF-IDF vectorizer from disk.
-    """
-    with open(path, "rb") as f:
-        vectorizer = pickle.load(f)
-
-    # Wrap into the embedding service
-    embedding_service = TFIDFEmbeddingService()
-    embedding_service.vectorizer = vectorizer
-    return embedding_service
 
 def run(db_session):
     try:
@@ -62,7 +49,7 @@ def run(db_session):
     
     # Embed job descriptions using SBERT and save to DB
     try:
-        embedding_service = SBERTEmbeddingService()
+        embedding_service = get_sbert_service()
 
         # Check for existing cluster embeddings to avoid duplicates
         existing_cluster_ids = {
@@ -91,6 +78,3 @@ def run(db_session):
 
     finally:
         db_session.close()
-
-if __name__ == "__main__":
-    main()
